requirements.txt
streamlit
redis
requests
pydantic
python-dotenv
flask
gunicorn
